boilerplate-projects/rag-chatbot/requirements.txt
streamlit>=1.29.0
langchain>=0.1.0
chromadb>=0.4.0
openai>=1.6.0
tiktoken>=0.5.0
pypdf>=3.17.0
python-dotenv>=1.0.0
